refactor: combine fermi download scripts - reduce source tree depth
